add response for comportement for dev

+++ app/src/Controller/UserController.php
@@ -18,7 +18,6 @@ use Symfony\Component\Routing\Annotation\Route;
 #[Route('/user')]
 class UserController extends AbstractController
 {
-
     use UserTrait;
 
     #[Route('/edit_account', name: 'app_user_edit_account', methods: ['GET', 'POST'])]
@@ -26,10 +25,10 @@ class UserController extends AbstractController
                             UserHelper $userHelper,
                             UserRepository $userRepository): Response
     {
-
         $user = $this->getUser();
 
-        $this->redirectIfNotAllow();
+        $response = $this->redirectIfNotAllow();
+        if($response) return $response;
 
         $session = $request->getSession();
         $session->set('previous_photo', $this->getUser()->getPhoto());
@@ -63,7 +62,8 @@ class UserController extends AbstractController
                             PaymentTransactionRepository $paymentTransactionRepository): Response
     {
 
-        $this->redirectIfNotAllow();
+        $response = $this->redirectIfNotAllow();
+        if($response) return $response;
 
         $dql = "SELECT SUM(p.amount) AS balance, p.payment_status, p.payment_for FROM App\Entity\PaymentTransaction p  WHERE p.payer = ?1 GROUP BY p.payment_for, p.payment_status";
         $paymentStats = $em->createQuery($dql)
@@ -96,7 +96,8 @@ class UserController extends AbstractController
     #[Route('/order', name: 'app_user_order', methods: ['GET'])]
     public function order(Request $request): Response
     {
-        $this->redirectIfNotAllow();
+        $response = $this->redirectIfNotAllow();
+        if($response) return $response;
 
         return $this->render('user/order.html.twig', [
             'user' => $this->getUser(),
@@ -106,7 +107,8 @@ class UserController extends AbstractController
     #[Route('/company', name: 'app_user_company', methods: ['GET'])]
     public function company(Request $request): Response
     {
-        $this->redirectIfNotAllow();
+        $response = $this->redirectIfNotAllow();
+        if($response) return $response;
 
         return $this->render('user/company.html.twig', [
             'user' => $this->getUser(),
@@ -116,7 +118,8 @@ class UserController extends AbstractController
     #[Route('/employee', name: 'app_user_employee', methods: ['GET'])]
     public function employee(Request $request): Response
     {
-        $this->redirectIfNotAllow();
+        $response = $this->redirectIfNotAllow();
+        if($response) return $response;
 
         return $this->render('user/employee.html.twig', [
             'user' => $this->getUser(),
@@ -126,7 +129,8 @@ class UserController extends AbstractController
     #[Route('/dashboard', name: 'app_user_dashboard', methods: ['GET'])]
     public function dashboard(Request $request): Response
     {
-        $this->redirectIfNotAllow();
+        $response = $this->redirectIfNotAllow();
+        if($response) return $response;
 
         return $this->render('user/dashboard.html.twig', [
             'user' => $this->getUser(),
@@ -136,7 +140,8 @@ class UserController extends AbstractController
     #[Route('/configuration', name: 'app_user_configuration', methods: ['GET'])]
     public function configuration(Request $request): Response
     {
-        $this->redirectIfNotAllow();
+        $response = $this->redirectIfNotAllow();
+        if($response) return $response;
 
         return $this->render('user/configuration.html.twig', [
             'user' => $this->getUser(),
@@ -146,7 +151,8 @@ class UserController extends AbstractController
     #[Route('/', name: 'app_user_index', methods: ['GET'])]
     public function index(Request $request, UserRepository $userRepository): Response
     {
-        $this->redirectIfNotAllow();
+        $response = $this->redirectIfNotAllow();
+        if($response) return $response;
 
         return $this->render('user/index.html.twig', [
             'users' => $userRepository->findAll(),
@@ -157,7 +163,8 @@ class UserController extends AbstractController
     #[Route('/{id}', name: 'app_user_show', methods: ['GET'])]
     public function show(Request $request, User $user): Response
     {
-        $this->redirectIfNotAllow();
+        $response = $this->redirectIfNotAllow();
+        if($response) return $response;
 
         return $this->render('user/show.html.twig', [
             'user' => $user,
@@ -167,7 +174,8 @@ class UserController extends AbstractController
     #[Route('/{id}/edit', name: 'app_user_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, User $user, UserRepository $userRepository): Response
     {
-        $this->redirectIfNotAllow();
+        $response = $this->redirectIfNotAllow();
+        if($response) return $response;
 
         $form = $this->createForm(UserFormType::class, $user);
         $form->handleRequest($request);
@@ -187,7 +195,8 @@ class UserController extends AbstractController
     #[Route('/{id}', name: 'app_user_delete', methods: ['POST'])]
     public function delete(Request $request, User $user, UserRepository $userRepository): Response
     {
-        $this->redirectIfNotAllow();
+        $response = $this->redirectIfNotAllow();
+        if($response) return $response;
 
         if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
             $userRepository->remove($user, true);
